feat: add typed `interfaces` config alongside legacy `endpoint`/`responsesEndpoint` #1615
### Applicable issues

- feat #1615

### Description of changes

Add a typed `interfaces` map on deployments (keyed by interface type, each with
a
`base_url`) as a **first-class peer** of the flat `endpoint` /
`responsesEndpoint`
fields — not a replacement. Both shapes are fully supported, per deployment,
per
interface type. **No migration**: no in-memory conversion, no on-disk rewrite,
no
settings flag.

- **Config**: add `InterfaceType` (`openaiChatCompletions`, `openaiResponses`)
  and
  `DeploymentInterface` (`base_url`); add `Deployment.interfaces` +
  helpers
  (`getInterfaceBaseUrl` / `resolveEndpoint` / `supportsInterface`);
  preserve
  `interfaces` in the `Application` copy constructor.
- **Routing**: dual-mode per interface type — `base_url` + the exact ingress
  path
  (e.g. `POST /openai/v1/responses` → `<base_url>/openai/v1/responses`), else
  the
  legacy endpoint **verbatim** as today. `interfaces` wins when both are
  declared.
  Wired into chat/completions, responses POST, responses GET/CANCEL/DELETE,
  and
  interceptors (with `{deployment-id}` segment rewrite).
- **Scope**: only `openaiChatCompletions` + `openaiResponses` are routed.
  Models
  support both and ignore unknown forward-compatible keys; applications
  and
  interceptors accept only `openaiChatCompletions` (others dropped on config
  read
  with a warning).
- **Backward compatible**: a deployment with only `endpoint` /
  `responsesEndpoint`
  routes byte-for-byte as today — no client/adapter change required.
- **Docs/tests**: document `interfaces` in
  `docs/dynamic-settings/{models,applications,interceptors}.md`
  and `sample/aidial.config.json`; add config, controller, and config-validation tests.

### Checklist

- [x] Title of the pull request
  follows [Conventional Commits specification](https://www.conventionalcommits.org/en/v1.0.0/)

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.
